added support for processing self & dot notation cal expressions & added several checks for nullables & illegal property access

+++ lib/runtime/symbols/Object.h
@@ -60,7 +60,7 @@ struct SharpObject
     void print();
 
     union {
-        double *HEAD;        /* data */
+        long double *HEAD;   /* data */
         Object *node;        /* structured data */
     };
